// src/enums/Constants.ts
export enum Constants {
    KINGSIDE_ROOK_COLUMN = 7,
    KINGSIDE_ROOK_DESTINATION_COLUMN = 5,
    KINGSIDE_CASTLING = 2,
    KINGSIDE_PASSED_SQUARE = 5,
    QUEENSIDE_PASSED_SQUARE = 3,
    QUEENSIDE_ROOK_COLUMN = 0,
    QUEENSIDE_ROOK_DESTINATION_COLUMN = 3,
    QUEENSIDE_CASTLING = -2,
    WHITE_END_ROW = 0,
    BLACK_END_ROW = 7
}